# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/file-api/v1/avatars": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Avatar"
                ],
                "summary": "上传头像",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.IDResult"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/file-api/v1/files": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "File"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.IDResult"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        },
        "/file-api/v1/files/{fileName}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "File"
                ],
                "summary": "下载指定文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "fileName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "File"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "唯一标识",
                        "name": "fileName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.StatusResult"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.ErrorResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "schema.ErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "schema.ErrorResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误项",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.ErrorItem"
                        }
                    ]
                }
            }
        },
        "schema.IDResult": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "schema.StatusResult": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "状态(OK)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/schema.StatusText"
                        }
                    ]
                }
            }
        },
        "schema.StatusText": {
            "type": "string",
            "enum": [
                "OK",
                "ERROR",
                "FAIL"
            ],
            "x-enum-varnames": [
                "OKStatus",
                "ErrorStatus",
                "FailStatus"
            ]
        }
    }
}